[all] removed dedicated .takeOver () methods and replace those with a generic sl::takeOver <T>
the absolute most of classes in axl can be taken over with a simple move semantics (destruct destination, memcpy, construct source), so having a generic non-member function is tempting; having a dedicated takeOver method is kind of a "cleaner" approach, but i think even better would be to clean up axl and make *everything* takeOver-compatible (and where it's not possible, explicitly disable takeOver with template-spec)

+++ src/axl_io/axl_io_Mapping.cpp
@@ -44,7 +44,7 @@ Mapping::open (
 	if (!result)
 		return NULL;
 
-	p = m_view.view (m_mapping, access, viewBegin, viewEnd - viewBegin);
+	p = m_view.view (m_mapping, access, viewBegin, (size_t) (viewEnd - viewBegin));
 	if (!p)
 	{
 		m_mapping.close ();
